refactor(registration-api): Rename SpringSecurityContextBean to SpringSecurityContextService and move into services package
