+++ applications/registration-api/src/main/java/no/dcat/service/SpringSecurityContextService.java
@@ -1,14 +1,11 @@
-package no.dcat.configuration;
+package no.dcat.service;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-/**
- * Spring bean to enable easier mocking of Spring Security static methods
- */
 @Service
-public class SpringSecurityContextBean {
+public class SpringSecurityContextService {
 
     public Authentication getAuthentication() {
         return SecurityContextHolder.getContext().getAuthentication();
